pyproject.toml
[tool.poetry]
name = "bet-maker"
version = "0.1.0"
description = ""
authors = ["MaxPositive <brunomax235@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.110.0"
redis = "^5.0.3"
pydantic-settings = "^2.2.1"
sqlalchemy = "^2.0.28"
uvicorn = "^0.28.0"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
alembic = "^1.13.1"
python-dotenv = "^1.0.1"


[tool.poetry.group.dev.dependencies]
black = "^24.2.0"
pytest = "^8.1.1"
pytest-asyncio = "^0.23.5.post1"
async-asgi-testclient = "^1.4.11"
pytest-env = "^1.1.3"
faker = "^24.1.0"
factory-boy = "^3.3.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110
